[refactor/#123] chat관련 파일
