+++ src/entities/chat/ui/index.ts
@@ -1,5 +1,4 @@
 export { ChatRoomItem } from "./ChatRoomItem";
-export { ChatRoomList } from "./ChatRoomList";
 export { MessageBubble } from "./MessageBubble";
 export { MessageItem } from "./MessageItem";
 export { MessageTime } from "./MessageTime";
