--- src/pages/partner/chat-room/ui/index.ts
@@ -1,1 +1,0 @@
-export { PartnerChatRoomPage } from "./PartnerChatRoomPage";
