+++ src/widgets/chat/chat-room-list/index.ts
@@ -1,0 +1,1 @@
+export { ChatRoomListWidget } from "./ui/ChatRoomListWidget";
