+++ src/widgets/chat/chat-room/ui/index.ts
@@ -1,4 +1,4 @@
 export { ChatHeader } from "./ChatHeader";
 export { ChatHeaderActions } from "./ChatHeaderActions";
 export { ChatMessageList } from "./ChatMessageList";
-export { ChatMessages } from "./ChatMessages";
+export { ChatRoomWidget } from "./ChatRoomWidget";
